+++ nodeconductor/cost_tracking/tasks.py
@@ -4,17 +4,19 @@ from nodeconductor.cost_tracking import CostTrackingRegister, models
 from nodeconductor.structure import models as structure_models
 
 
-@shared_task(name='nodeconductor.cost_tracking.recalculate_consumed_estimate')
-def recalculate_consumed_estimate():
-    """ Recalculate how many consumables were used by resource until now.
+@shared_task(name='nodeconductor.cost_tracking.recalculate_estimate')
+def recalculate_estimate(recalculate_total=False):
+    """ Recalculate price of consumables that were used by resource until now.
 
         Regular task. It is too expensive to calculate consumed price on each
         request, so we store cached price each hour.
+        If recalculate_total is True - task also recalculates total estimate
+        for current month.
     """
     # Step 1. Recalculate resources estimates.
     for resource_model in CostTrackingRegister.registered_resources:
         for resource in resource_model.objects.all():
-            _update_resource_consumed(resource)
+            _update_resource_consumed(resource, recalculate_total=recalculate_total)
     # Step 2. Move from down to top and recalculate consumed estimate for each
     #         object based on its children.
     ancestors_models = [m for m in models.PriceEstimate.get_estimated_models()
@@ -24,11 +26,13 @@ def recalculate_consumed_estimate():
             _update_ancestor_consumed(ancestor)
 
 
-def _update_resource_consumed(resource):
+def _update_resource_consumed(resource, recalculate_total):
     price_estimate, created = models.PriceEstimate.objects.get_or_create_current(scope=resource)
     if created:
         models.ConsumptionDetails.objects.create(price_estimate=price_estimate)
         price_estimate.create_ancestors()
+        price_estimate.update_total()
+    elif recalculate_total:
         price_estimate.update_total()
     price_estimate.update_consumed()